pyproject.toml
[project]
name = "sv-comp-helper"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "typer>=0.15.2",
]